requirements.txt
flask
openai
python-dotenv
pdf2image
Werkzeug 